test-requirements.txt
djangorestframework==3.6.3
flake8==3.3.0
psycopg2==2.7.1
pylint==1.7.2
pytest-cov==2.5.1
pytest-env==0.6.2
pytest==3.1.3
six==1.10.0
tox-travis==0.8
tox==2.7.0
typing==3.6.1
